codacy correction
