+++ app/src/main/AndroidManifest.xml
@@ -62,6 +62,11 @@
         <activity
             android:name="com.alfresco.content.browse.preview.PreviewActivity"
             android:configChanges="orientation|screenSize" />
+
+        <activity
+            android:name=".activity.MoveActivity"
+            android:windowSoftInputMode="adjustNothing" />
+
     </application>
 
 </manifest>
